requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.10
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
